+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added dependency "[cake-case](https://github.com/archco/cake-case#readme)". #103
 - Added style for `<kbd>` tag. #109
 - Added scss-class `.btn-text`. #113
+- Added a new vue-directive `v-outer`. #112
 
 ### Changed